VRUI : UI / Manager 분리 작업 1) UI: Presentation 2) VRManager : Domain
